attempting to call context_switch assembly
