+++ CMakeLists.txt
@@ -37,7 +37,7 @@ set(CMAKE_CXX_FLAGS_DEBUG "-g -O0 -Wall")
 set(CMAKE_CXX_FLAGS_RELEASE "-O0")
 # Add executable. Default name is the project name, version 0.1
 
-add_executable(asquaredos asquaredos.c)
+add_executable(asquaredos asquaredos.c kern/context_switch.s)
 
 pico_set_program_name(asquaredos "asquaredos")
 pico_set_program_version(asquaredos "0.1")
