+++ kern/palloc.h
@@ -7,24 +7,7 @@
 #define __PALLOC_H__
 
 #include "resources.h"
-
-/*
- * Represents a single region of memory that has been allocated to, or could
- * be allocated to a process, by the kernel.
- */
-typedef struct heap_region {
-    uint32_t size;   /* Size of region in bytes. */
-
-    /*
-     * Multiplexed between global free list when not allocated, and the kernel's
-     * private accounting of a process's allocated list when the region has been
-     * allocated.
-     */
-    struct heap_region     *next;
-    struct heap_region     *prev;
-
-    uint8_t data[];     /* Beginning of region's user data. */
-} heap_region_t;
+#include "scheduler.h"
 
 /*
  * Allocate memory for a userspace process.
