+++ kern/scheduler.h
@@ -6,12 +6,28 @@
 #ifndef __SCHED_H__
 #define __SCHED_H__
 
-#include "palloc.h"
-
 /*
  * 32-bit register value.
  */
 typedef uint32_t register_t;
+
+/*
+ * Represents a single region of memory that has been allocated to, or could
+ * be allocated to a process, by the kernel.
+ */
+typedef struct heap_region {
+    uint32_t size;   /* Size of region in bytes. */
+
+    /*
+     * Multiplexed between global free list when not allocated, and the kernel's
+     * private accounting of a process's allocated list when the region has been
+     * allocated.
+     */
+    struct heap_region     *next;
+    struct heap_region     *prev;
+
+    uint8_t data[];     /* Beginning of region's user data. */
+} heap_region_t;
 
 /*
  * The format in which registers are pushed to the stack during a context
